feat: add graphql basic data (#13)
* install graphql

* wrap app in graphql provider.

* change loader components

* add some spinners components/ls

* rm schema

* rfm schema
