--- app/(auth)/register/components/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-# Keep environment variables out of version control
- sc.env
