+++ app/(base)/(misc)/blackjack/page.tsx
@@ -3,7 +3,7 @@ import React, { useState, useEffect } from 'react';
 import { Card } from '@/components/ui/card';
 import { Input } from '@/components/ui/input';
 import { Button } from '@/components/ui/button';
-import blackjackData from '@/lib/blackjack.json';
+import blackjackData from '@/lib/config/blackjack.json';
 import { toast } from 'sonner'
 import SingleHand from '@/components/misc/BlackjackSingleHand';
 import { Accordion, AccordionContent, AccordionItem, AccordionTrigger } from '@/components/ui/accordion';
