+++ app/(base)/(misc)/transactions/page.tsx
@@ -19,7 +19,7 @@ import BalanceDisplay from "@/components/transactions/Balance";
 import TransactionForm from "@/components/transactions/TransactionForm";
 import TransactionList from "@/components/transactions/TransactionList";
 import { toast } from 'sonner';
-import { db } from "@/lib/firebase";
+import { db } from "@/lib/firebase/firebase";
 
 const TransactionPage = () => {
     const [transactions, setTransactions] = useState<Transaction[]>([]);
@@ -33,7 +33,7 @@ const TransactionPage = () => {
     useEffect(() => {
         let unsubscribe: Unsubscribe;
         const loadDb = async () => {
-            const module = await import('@/lib/firebase');
+            const module = await import('@/lib/firebase/firebase');
             return module.db;
         };
