--- docker-compose.yml
@@ -1,10 +1,0 @@
-services:
-  # ...keep existing 'backend' service
-
-  postgres:
-    image: postgres:15-alpine
-    ports:
-      - 5432:5432  # enables connecting locally via psql
-    environment:
-      POSTGRES_DB: webapp_dev
-      POSTGRES_HOST_AUTH_METHOD: trust
