+++ lib/(graphql)/ApolloClient.ts
@@ -1,0 +1,12 @@
+import { ApolloClient, InMemoryCache } from '@apollo/client';
+
+const client = new ApolloClient({
+    uri: 'https://api.github.com/graphql',
+    cache: new InMemoryCache(),
+    headers: {
+        Authorization: `Bearer ${process.env.GITHUB_TOKEN}`
+    },
+});
+
+export default client;
+
