--- lib/blackjack.json
@@ -1,364 +1,0 @@
-{
-  "strategy": {
-    "2": {
-      "2": "HIT",
-      "3": "HIT",
-      "4": "HIT",
-      "5": "HIT",
-      "6": "HIT",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "3": {
-      "2": "HIT",
-      "3": "HIT",
-      "4": "HIT",
-      "5": "HIT",
-      "6": "HIT",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "4": {
-      "2": "HIT",
-      "3": "HIT",
-      "4": "HIT",
-      "5": "HIT",
-      "6": "HIT",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "5": {
-      "2": "HIT",
-      "3": "HIT",
-      "4": "HIT",
-      "5": "HIT",
-      "6": "HIT",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "6": {
-      "2": "HIT",
-      "3": "HIT",
-      "4": "HIT",
-      "5": "HIT",
-      "6": "HIT",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "7": {
-      "2": "HIT",
-      "3": "HIT",
-      "4": "HIT",
-      "5": "HIT",
-      "6": "HIT",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "8": {
-      "2": "HIT",
-      "3": "HIT",
-      "4": "HIT",
-      "5": "HIT",
-      "6": "HIT",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "9": {
-      "2": "HIT",
-      "3": "Dubble",
-      "4": "Dubble",
-      "5": "Dubble",
-      "6": "Dubble",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "10": {
-      "2": "Dubble",
-      "3": "Dubble",
-      "4": "Dubble",
-      "5": "Dubble",
-      "6": "Dubble",
-      "7": "Dubble",
-      "8": "Dubble",
-      "9": "Dubble",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "11": {
-      "2": "Dubble",
-      "3": "Dubble",
-      "4": "Dubble",
-      "5": "Dubble",
-      "6": "Dubble",
-      "7": "Dubble",
-      "8": "Dubble",
-      "9": "Dubble",
-      "10": "Dubble",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "12": {
-      "2": "HIT",
-      "3": "HIT",
-      "4": "Stay",
-      "5": "Stay",
-      "6": "Stay",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "13": {
-      "2": "Stay",
-      "3": "Stay",
-      "4": "Stay",
-      "5": "Stay",
-      "6": "Stay",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "14": {
-      "2": "Stay",
-      "3": "Stay",
-      "4": "Stay",
-      "5": "Stay",
-      "6": "Stay",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "15": {
-      "2": "Stay",
-      "3": "Stay",
-      "4": "Stay",
-      "5": "Stay",
-      "6": "Stay",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "16": {
-      "2": "Stay",
-      "3": "Stay",
-      "4": "Stay",
-      "5": "Stay",
-      "6": "Stay",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "17": {
-      "2": "Stay",
-      "3": "Stay",
-      "4": "Stay",
-      "5": "Stay",
-      "6": "Stay",
-      "7": "Stay",
-      "8": "Stay",
-      "9": "Stay",
-      "10": "Stay",
-      "A": "Stay",
-      "J": "Stay",
-      "K": "Stay",
-      "Q": "Stay"
-    },
-    "A2": {
-      "2": "HIT",
-      "3": "HIT",
-      "4": "HIT",
-      "5": "Dubble",
-      "6": "Dubble",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "A3": {
-      "2": "HIT",
-      "3": "HIT",
-      "4": "HIT",
-      "5": "Dubble",
-      "6": "Dubble",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "A4": {
-      "2": "HIT",
-      "3": "HIT",
-      "4": "Dubble",
-      "5": "Dubble",
-      "6": "Dubble",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "A5": {
-      "2": "HIT",
-      "3": "HIT",
-      "4": "Dubble",
-      "5": "Dubble",
-      "6": "Dubble",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "A6": {
-      "2": "HIT",
-      "3": "Dubble",
-      "4": "Dubble",
-      "5": "Dubble",
-      "6": "Dubble",
-      "7": "HIT",
-      "8": "HIT",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "A7": {
-      "2": "Dubble",
-      "3": "Dubble",
-      "4": "Dubble",
-      "5": "Dubble",
-      "6": "Dubble",
-      "7": "Stay",
-      "8": "Stay",
-      "9": "HIT",
-      "10": "HIT",
-      "A": "HIT",
-      "J": "HIT",
-      "K": "HIT",
-      "Q": "HIT"
-    },
-    "A8": {
-      "2": "Stay",
-      "3": "Stay",
-      "4": "Stay",
-      "5": "Stay",
-      "6": "Stay",
-      "7": "Stay",
-      "8": "Stay",
-      "9": "Stay",
-      "10": "Stay",
-      "A": "Stay",
-      "J": "Stay",
-      "K": "Stay",
-      "Q": "Stay"
-    },
-    "A9": {
-      "2": "Stay",
-      "3": "Stay",
-      "4": "Stay",
-      "5": "Stay",
-      "6": "Stay",
-      "7": "Stay",
-      "8": "Stay",
-      "9": "Stay",
-      "10": "Stay",
-      "A": "Stay",
-      "J": "Stay",
-      "K": "Stay",
-      "Q": "Stay"
-    }
-  }
-}
