+++ lib/data/graphql.md
@@ -1,0 +1,19 @@
+# Steps for graphql test.
+
+-Create a connection (lib/apolloclient)
+      - backend api keys or in my case github access token
+
+- Wrap project in apollo provider.
+
+- Write querys
+
+
+
+## mutation
+ Mutation is een type query.
+ Mutatie -> veranderen
+    <api>
+
+
+- Iets van data muteren
+- Vervolgens die data real time laten updaten d.m.v subscriptions)
