+++ package.json
@@ -9,7 +9,10 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@apollo/client": "^3.8.7",
+    "@apollo/experimental-nextjs-app-support": "^0.5.1",
     "@formkit/auto-animate": "^0.8.1",
+    "@next/font": "^14.0.3",
     "@nextui-org/react": "^2.2.0",
     "@prisma/client": "^5.5.2",
     "@radix-ui/react-accordion": "^1.1.2",
@@ -24,6 +27,7 @@
     "firebase": "^10.6.0",
     "framer-motion": "^10.16.4",
     "geist": "^1.0.0",
+    "graphql": "^16.8.1",
     "lucide-react": "^0.288.0",
     "next": "13.5.6",
     "prisma": "^5.5.2",
