+++ styles/components/ ./spinner-cutout.scss
@@ -1,0 +1,44 @@
+.spinner-cutout {
+    display: block;
+    z-index: 99;
+    margin-top: 40vh;
+    margin-left: auto;
+    margin-right: auto;
+    width: 100px;
+    height: 100px;
+    border: 3px solid white;
+    border-radius: 50%;
+
+    animation: spin 7s ease-in-out;
+    animation-iteration-count: infinite;
+    transition-duration: 0.1s;
+}
+
+.spinner-cutout:hover {
+    scale: 0.95;
+}
+
+.spinner-cutout:active {
+    scale: 2.5;
+}
+
+@keyframes spin {
+    0% {
+        transform: rotate(0deg);
+        border-bottom: solid 3px transparent;
+        border-top: solid 3px transparent;
+    }
+
+    50% {
+        transform: rotate(1800deg);
+        border: 3px solid white;
+        border-left: solid 3px transparent;
+        border-right: solid 3px transparent;
+    }
+
+    100% {
+        transform: rotate(0deg);
+        border-bottom: solid 3px transparent;
+        border-top: solid 3px transparent;
+    }
+}
